remove lib folder
