+++ .travis.yml
@@ -1,4 +1,5 @@
 sudo: false
+os: linux
 language: node_js
 node_js:
   - "0.10"
@@ -7,10 +8,10 @@ node_js:
   - "io.js"
   - "4"
   - "5"
+env:
+  - TEST_SUITE=unit
 matrix:
   include:
     - node_js: "4"
       env: TEST_SUITE=lint
-env:
-  - TEST_SUITE=unit
 script: npm run-script $TEST_SUITE
