+++ types/bound.js
