+++ types/buffer.js
