+++ types/varstring.js
